requirements.txt
Django==3.0.10
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
pytest-django==4.1.0
